test: extract toolbar buttons tests into separate suite
