+++ packages/rich-text-editor/test/toolbar-lit.test.js
@@ -1,0 +1,2 @@
+import '../theme/lumo/vaadin-lit-rich-text-editor.js';
+import './basic.common.js';
