+++ packages/rich-text-editor/test/toolbar-polymer.test.js
@@ -1,0 +1,2 @@
+import '../theme/lumo/vaadin-rich-text-editor.js';
+import './basic.common.js';
